# packages/schemas/recipe/openapi.yaml
...
openapi: 3.0.3
info:
  title: Commise Recipe API
  version: 1.0.0
  description: |-
    The recipe service owns this contract. It is DERIVED from the service’s authored zod
    (`src/**/*.schema.ts` plus the `@kitchensink/recipe-core` schemas those compose), never
    hand-written, and it is not a code-generation input.

    Every path is also served under a deprecated bare `/v1/...` alias for consumers configured
    outside this repo and for already-shipped mobile binaries (ADR-0011). New integrators must use
    the `/api/v1/...` spelling published here.

    Coverage is partial and the gaps are enumerated by the generator on every run: an operation with
    no response schema is one whose body has no authored zod yet, not one whose body is empty.
servers:
  - url: https://recipe.commise.app
    description: Production
  - url: https://recipe.sandbox.commise.app
    description: Sandbox
security:
  - clerkSessionToken: []
paths:
...
  /api/v1/collections:
    post:
      operationId: createCollection
      summary: Create a collection owned by the caller.
      security:
        - clerkSessionToken: []
      responses:
        "201":
          description: The created collection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The caller is at the 50-collection cap.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      description: Visibility defaults to `private` (FR-010). `ownerId` is not an accepted field — ownership comes from the
        verified token, and the body is strict, so an `ownerId` is rejected with a 400 rather than stripped.
      requestBody:
        description: The collection to create.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollectionRequest"
    get:
      operationId: listCollections
      summary: List the caller’s collections.
      security:
        - clerkSessionToken: []
      responses:
        "200":
          description: A page of collections.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCollections"
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      description: Newest first. `recipeCount` is OMITTED on this list projection (it is present on the single-collection
        read) — counting members per row would be an N+1.
      parameters:
        - name: page
          in: query
          required: false
          description: 1-based page number.
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          required: false
          description: Page size, 1..100.
          schema:
            type: integer
            minimum: 1
            maximum: 100
  /api/v1/collections/{id}:
    get:
      operationId: getCollectionById
      summary: Read a collection with its member recipes.
      security:
        - clerkSessionToken: []
      responses:
        "200":
          description: The collection and its members.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionWithRecipes"
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such collection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      description: "`recipes` is ALWAYS present — an empty array means “no members you can see”, which is a real state.
        Members are filtered to those the caller may view, so a member that has since gone private simply disappears.
        `coverPhotoUrl` is never populated on this embed."
      parameters:
        - name: id
          in: path
          required: true
          description: The collection id.
          schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
    patch:
      operationId: updateCollection
      summary: Update a collection (at least one field required).
      security:
        - clerkSessionToken: []
      responses:
        "200":
          description: The updated collection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: The caller does not own the resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such collection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      description: "AT LEAST ONE FIELD IS REQUIRED, and that rule is NOT expressible in this document: it is a zod
        `.refine()`, which JSON Schema cannot represent, so an empty object validates here and is answered `400` by the
        service. Stated rather than dropped — dropping the rule would turn a client bug into a silent success."
      parameters:
        - name: id
          in: path
          required: true
          description: The collection id.
          schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
      requestBody:
        description: The fields to change.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCollectionRequest"
    delete:
      operationId: deleteCollection
      summary: Delete a collection (its recipes are untouched).
      security:
        - clerkSessionToken: []
      responses:
        "204":
          description: Deleted.
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: The caller does not own the resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such collection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: id
          in: path
          required: true
          description: The collection id.
          schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
  /api/v1/collections/{id}/recipes:
    post:
      operationId: addRecipeToCollection
      summary: Add a recipe to a collection.
      security:
        - clerkSessionToken: []
      responses:
        "201":
          description: The created membership.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionRecipeMembership"
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: The caller does not own the resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such collection or recipe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      description: Idempotent. A recipe the caller cannot VIEW is answered `404`, not `403`, so a private recipe’s existence
        is never disclosed. The response body’s timestamp field is `createdAt`, while the persisted column and the
        shared domain type both call it `addedAt`.
      parameters:
        - name: id
          in: path
          required: true
          description: The collection id.
          schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
      requestBody:
        description: The recipe to add.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRecipeToCollectionRequest"
  /api/v1/collections/{id}/recipes/{recipeId}:
    delete:
      operationId: removeRecipeFromCollection
      summary: Remove a recipe from a collection.
      security:
        - clerkSessionToken: []
      responses:
        "204":
          description: Removed.
        "400":
          description: "The request body, query or path was rejected. `code` says by what: `VALIDATION_FAILED` from the boundary
            parser (with `details.fields` naming each offending field, including a rejected unknown key), or a domain
            code such as `INVALID_VISIBILITY` / `UNKNOWN_INGREDIENT`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No, invalid, expired, or wrong-authorized-party bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: The caller does not own the resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such collection or membership.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: "The caller has an in-flight account erasure, so writes are locked (HAZ-052). `code: ERASURE_IN_PROGRESS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: "Per-user rate limit exceeded (`code: TOO_MANY_REQUESTS`). ⚠️ This body used to be a bare JSON STRING — the
            throttler exception passed through the filter unchanged — and is now the envelope."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          headers:
            Retry-After:
              description: Seconds until the limit resets.
              schema:
                type: integer
                minimum: 0
                exclusiveMinimum: true
        "500":
          description: Unexpected server error. The body never carries internals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: id
          in: path
          required: true
          description: The collection id.
          schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        - name: recipeId
          in: path
          required: true
          description: The recipe to remove.
          schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
...
components:
  securitySchemes:
    clerkSessionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A Clerk session token. Verified networklessly against the instance JWT key, with the signed `azp` checked
        against the service’s authorized-party pattern.
    serviceErasureToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A signed service-principal token, used only by the internal erasure route.
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: {}
      required:
        - code
        - message
      additionalProperties: {}
...
    Collection:
      type: object
      properties:
        id:
          type: string
          minLength: 1
        ownerId:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        visibility:
          type: string
          enum:
            - public
            - private
        description:
          type: string
          minLength: 1
        sourceCollectionId:
          type: string
          minLength: 1
        createdAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
        updatedAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
        recipeCount:
          type: integer
          minimum: 0
        sourceOwnerHandle:
          type: string
          minLength: 1
        sourceCollectionName:
          type: string
          minLength: 1
        lastPulledAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
      required:
        - id
        - ownerId
        - name
        - visibility
        - createdAt
        - updatedAt
    PaginatedCollections:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Collection"
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 0
          exclusiveMinimum: true
        pageSize:
          type: integer
          minimum: 0
          exclusiveMinimum: true
        hasMore:
          type: boolean
      required:
        - data
        - total
        - page
        - pageSize
        - hasMore
...
    CreateCollectionRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          minLength: 1
          maxLength: 1000
        visibility:
          type: string
          enum:
            - public
            - private
      required:
        - name
      additionalProperties: false
    UpdateCollectionRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          minLength: 1
          maxLength: 1000
        visibility:
          type: string
          enum:
            - public
            - private
      additionalProperties: false
    AddRecipeToCollectionRequest:
      type: object
      properties:
        recipeId:
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
      required:
        - recipeId
      additionalProperties: false
...